+++ lib/model/location_result.dart
@@ -5,16 +5,15 @@ class LocationResult {
   /// The human readable name of the location. This is primarily the
   /// name of the road. But in cases where the place was selected from Nearby
   /// places list, we use the <b>name</b> provided on the list item.
-  String name; // or road
-
-  /// The human readable locality of the location.
-  String locality;
+  String address; // or road
 
   /// Latitude/Longitude of the selected location.
   LatLng latLng;
 
+  LocationResult({this.latLng, this.address});
+
   @override
   String toString() {
-    return 'LocationResult{name: $name, locality: $locality, latLng: $latLng}';
+    return 'LocationResult{address: $address, latLng: $latLng}';
   }
 }