+++ lib/model/nearby_place.dart
@@ -3,7 +3,7 @@ import 'package:google_maps_flutter/google_maps_flutter.dart';
 /// Nearby place data will be deserialized into this model.
 class NearbyPlace {
   /// The human-readable name of the location provided. This value is provided
-  /// for [LocationResult.name] when the user selects this nearby place.
+  /// for [LocationResult.address] when the user selects this nearby place.
   String name;
 
   /// The icon identifying the kind of place provided. Eg. lodging, chapel,